requirements.txt
opencv-python
google-cloud-vision
celery==4.1.1
falcon==1.4.1
flower==0.9.2
gunicorn==19.7.1
redis==2.10.6
nltk==3.4.5
pyenchant==2.0.0
pandas
imutils
fuzzywuzzy
celery-with-mongodb
setuptools
google-cloud-vision==1.0.0
PyPDF2
tabula
textblob
wand
pillow
pikepdf
wheel
pdf2image
pdf2text
flask
PyJWT
pymongo
flask_cors
regex
xlwt
xlrd
xlsxwriter
pdftotext
pytesseract
tensorflow
scikit-learn==0.20.3
scikit-image
scipy
seaborn
openpyxl
default-jre
openjdk-11-jre-headless
openjdk-8-jre-headless